refactor: extract the validation error handling into a separate module
