+++ backend/validation/adminValidation.js
@@ -1,23 +1,11 @@
-import { body, validationResult } from "express-validator";
-import { sendError } from "../utils/response.js";
+import { body } from "express-validator";
+import { validationError } from "./validationError.js";
 
 class AdminValidation {
   checkUserId = [
     body("userId").trim().notEmpty().withMessage("User ID is required"),
 
-    (req, res, next) => {
-      const errors = validationResult(req);
-      if (!errors.isEmpty()) {
-        return sendError(res, {
-          status: 400,
-          message: `Validation failed: ${errors
-            .array()
-            .map((entry) => `${entry.path}: ${entry.msg}`)
-            .join(", ")}.`,
-        });
-      }
-      next();
-    },
+    validationError,
   ];
 }
