+++ backend/validation/authValidation.js
@@ -1,6 +1,6 @@
-import { body, validationResult, param } from "express-validator";
+import { body, param } from "express-validator";
 import { usersModel } from "../models/usersModel.js";
-import { sendError } from "../utils/response.js";
+import { validationError } from "./validationError.js";
 
 class AuthValidation {
   signupValidationRules = [
@@ -66,54 +66,18 @@ class AuthValidation {
         return true;
       }),
 
-    (req, res, next) => {
-      const errors = validationResult(req);
-      if (!errors.isEmpty()) {
-        return sendError(res, {
-          status: 400,
-          message: `Validation failed: ${errors
-            .array()
-            .map((entry) => `${entry.path}: ${entry.msg}`)
-            .join(", ")}.`,
-        });
-      }
-      next();
-    },
+    validationError,
   ];
 
   confirmTokenValidationRules = [
     param("token").trim().notEmpty().withMessage("Token is required"),
-    (req, res, next) => {
-      const errors = validationResult(req);
-      if (!errors.isEmpty()) {
-        return sendError(res, {
-          status: 400,
-          message: `Validation failed: ${errors
-            .array()
-            .map((entry) => `${entry.path}: ${entry.msg}`)
-            .join(", ")}.`,
-        });
-      }
-      next();
-    },
+    validationError,
   ];
 
   loginValidationRules = [
     body("username").trim().notEmpty().withMessage("Username is required"),
     body("password").trim().notEmpty().withMessage("Password is required"),
-    (req, res, next) => {
-      const errors = validationResult(req);
-      if (!errors.isEmpty()) {
-        return sendError(res, {
-          status: 400,
-          message: `Validation failed: ${errors
-            .array()
-            .map((entry) => `${entry.path}: ${entry.msg}`)
-            .join(", ")}.`,
-        });
-      }
-      next();
-    },
+    validationError,
   ];
 }
